restructure into pages folder
